some fixes to tempcrap
