+++ sql/src/test/java/org/apache/druid/sql/calcite/util/CalciteTestBase.java
@@ -30,11 +30,22 @@ import org.apache.druid.server.security.ResourceType;
 import org.apache.druid.sql.calcite.expression.DruidExpression;
 import org.apache.druid.sql.calcite.expression.SimpleExtraction;
 import org.apache.druid.sql.http.SqlParameter;
+import org.junit.Rule;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInfo;
+import org.junit.jupiter.api.io.TempDir;
+import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
+import org.junit.rules.TemporaryFolder;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.Collections;
 import java.util.List;
 
+@EnableRuleMigrationSupport
 public abstract class CalciteTestBase
 {
   public static final List<SqlParameter> DEFAULT_PARAMETERS = ImmutableList.of();
@@ -45,6 +56,69 @@ public abstract class CalciteTestBase
     NullHandling.initializeForTests();
     ExpressionProcessing.initializeForTests();
   }
+
+  @Rule(order = 2)
+  public static TemporaryFolder temporaryFolder2 = new TemporaryFolder();
+
+  /**
+   * Temporary folder(s).
+   *
+   * Due to some possible reuse of configuration in the next case; there is only one real temporary path;
+   * but every case gets a separate folder in it.
+   *
+   * note: {@link #rootTempPath} and {@link #casetempPath} are made private to ensure that
+   */
+  @TempDir
+  private static Path rootTempPath;
+  private Path casetempPath;
+
+  @BeforeEach
+  public void setCaseTempDir(TestInfo testInfo) throws IOException
+  {
+    String methodName = testInfo.getTestMethod().get().getName();
+    casetempPath = Files.createTempDirectory(rootTempPath, methodName);
+  }
+
+  public File newTempFolder(String prefix)
+  {
+    try {
+      return Files.createTempDirectory(casetempPath, prefix).toFile();
+    }
+    catch (IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  public File newTempFile(String prefix)
+  {
+    try {
+      return Files.createTempFile(casetempPath, prefix, null).toFile();
+    }
+    catch (IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  public TempFolderOverTempDir temporaryFolder = new TempFolderOverTempDir();
+
+  public class TempFolderOverTempDir {
+
+    public File newFile()
+    {
+      return newTempFile("tempFile");
+    }
+
+    public File newFolder()
+    {
+      return newTempFolder("unknown");
+    }
+
+    public File newFolder(String string)
+    {
+      return newTempFolder(string);
+    }
+  }
+
 
   /**
    * @deprecated prefer to make {@link DruidExpression} directly to ensure expression tests accurately test the full
