+++ sql/src/test/java/org/apache/druid/sql/http/SqlResourceTest.java
@@ -105,15 +105,13 @@ import org.apache.druid.sql.calcite.util.CalciteTests;
 import org.apache.druid.sql.calcite.util.QueryLogHook;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
-import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.rules.TemporaryFolder;
-
+import org.junit.jupiter.api.io.TempDir;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.servlet.http.HttpServletRequest;
@@ -122,6 +120,7 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.StreamingOutput;
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.AbstractList;
@@ -169,8 +168,7 @@ public class SqlResourceTest extends CalciteTestBase
 
   private static Closer staticCloser = Closer.create();
   private static QueryRunnerFactoryConglomerate conglomerate;
-  @ClassRule
-  public static TemporaryFolder temporaryFolder = new TemporaryFolder();
+
   private static SpecificSegmentsQuerySegmentWalker walker;
   private static QueryScheduler scheduler;
 
@@ -198,7 +196,7 @@ public class SqlResourceTest extends CalciteTestBase
   private static final AtomicReference<Supplier<Void>> SCHEDULER_BAGGAGE = new AtomicReference<>();
 
   @BeforeAll
-  public static void setupClass() throws Exception
+  public static void setupClass(@TempDir File tempDir) throws Exception
   {
     conglomerate = QueryStackTests.createQueryRunnerFactoryConglomerate(staticCloser);
     scheduler = new QueryScheduler(
@@ -221,7 +219,7 @@ public class SqlResourceTest extends CalciteTestBase
         );
       }
     };
-    walker = CalciteTests.createMockWalker(conglomerate, temporaryFolder.newFolder(), scheduler);
+    walker = CalciteTests.createMockWalker(conglomerate, tempDir, scheduler);
     staticCloser.register(walker);
   }
